OneFragmentの変更
