+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ItemDiffUtil.kt
@@ -1,0 +1,13 @@
+package jp.co.yumemi.android.code_check
+
+import androidx.recyclerview.widget.DiffUtil
+
+object ItemDiffUtil : DiffUtil.ItemCallback<Item>() {
+    override fun areItemsTheSame(oldItem: Item, newItem: Item): Boolean {
+        return oldItem.name == newItem.name
+    }
+
+    override fun areContentsTheSame(oldItem: Item, newItem: Item): Boolean {
+        return oldItem == newItem
+    }
+}
