+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ListViewFragment.kt
@@ -1,0 +1,69 @@
+/*
+ * Copyright © 2021 YUMEMI Inc. All rights reserved.
+ */
+package jp.co.yumemi.android.code_check
+
+import android.os.Bundle
+import android.view.View
+import android.view.inputmethod.EditorInfo
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.viewModels
+import androidx.lifecycle.lifecycleScope
+import androidx.navigation.fragment.findNavController
+import androidx.recyclerview.widget.*
+import jp.co.yumemi.android.code_check.databinding.FragmentOneBinding
+import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.flow.map
+
+class ListViewFragment : Fragment(R.layout.fragment_one) {
+
+    private val viewModel by viewModels<ListViewModel> {
+        OneViewModelFactory()
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        val binding = FragmentOneBinding.bind(view)
+        val layoutManager = LinearLayoutManager(requireContext())
+        val dividerItemDecoration =
+            DividerItemDecoration(requireContext(), layoutManager.orientation)
+        val adapter = CustomAdapter(object : CustomAdapter.OnItemClickListener {
+            override fun itemClick(item: Item) {
+                gotoRepositoryFragment(item)
+            }
+        })
+
+        binding.searchInputText.setOnEditorActionListener { editText, action, _ ->
+            if (action == EditorInfo.IME_ACTION_SEARCH) {
+                viewModel.searchRepository(editText.text.toString())
+                return@setOnEditorActionListener true
+            }
+            return@setOnEditorActionListener false
+        }
+        lifecycleScope.launchWhenStarted {
+            viewModel.searchResults.map {
+                it.map { item ->
+                    item.copy(
+                        language = requireContext().getString(R.string.written_language, item.language)
+                    )
+                }
+            }.collect {
+                adapter.submitList(it)
+            }
+        }
+
+        binding.recyclerView.also {
+            it.layoutManager = layoutManager
+            it.addItemDecoration(dividerItemDecoration)
+            it.adapter = adapter
+        }
+    }
+
+    fun gotoRepositoryFragment(item: Item) {
+        val action = ListViewFragmentDirections
+            .actionRepositoriesFragmentToRepositoryFragment(item = item)
+        findNavController().navigate(action)
+    }
+}
+
