+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/oneFragment"
-        android:name="jp.co.yumemi.android.code_check.OneFragment"
+        android:name="jp.co.yumemi.android.code_check.ListViewFragment"
         android:label="@string/app_name"
         tools:layout="@layout/fragment_one">
         <action
